Add solution for 2023 - Day 22 Part 1
This day was really annoying as I was getting close to the answer,
but this slightly off. After a lot of debugging, visualisation and
comparing output with another solution I finally figured it out.
It was my use of `itertools::Itertools::at_most_one` as it will quit
early it if the condition aren't met. This was a problem as I was
calling HashMap::insert to update the height map as part of the
iterator, and quitting early made it so the height map was only
partially updates as a result. Super annoying. :(

/rant