Add module common_v2
The addition of module "common_v2" is to create a new way to write
solutions for each day. It attempts fix and improve previous
version. Like allowing part1 and part2 to run indepedent, so that
errors in part2 does not cause part1 to also fail like it does is
v1. But also it allows them to share code, like parsing or building
a common data-structure, instead of having to functions that can
lead to alot of duplicate code.

I am also hoping to create better error handling and CLI interface.